+++ storage/storage.go
@@ -57,12 +57,12 @@ type Storage interface {
 
 // NewClient returns new Storage client from given url. Storage implementation
 // is inferred from the url.
-func NewClient(url *url.URL) Storage {
+func NewClient(url *url.URL, storageOpts StorageOptions) (Storage, error) {
 	if url.IsRemote() {
-		return cachedS3
+		return NewS3Storage(storageOpts)
 	}
 
-	return NewFilesystem()
+	return NewFilesystem(), nil
 }
 
 // Object is a generic type which contains metadata for storage items.
@@ -198,4 +198,14 @@ type notImplemented struct {
 // Error returns the string representation of Error for notImplemented.
 func (e notImplemented) Error() string {
 	return fmt.Sprintf("%q is not supported on %q storage", e.method, e.apiType)
+}
+
+type StorageOptions struct {
+	NoVerifySSL       bool
+	Concurrency       int
+	MaxRetries        int
+	PartSize          int64
+	SourceRegion      string
+	DestinationRegion string
+	Endpoint          string
 }